+++ cmd/sparkwing/main.go
@@ -425,8 +425,6 @@ func runSparkwing(args []string) error {
 		return dispatchRun(args)
 	case "wingd":
 		return runWingd(args[1:])
-	case queueExecGuardCommandName:
-		return runQueueExecGuard(args[1:])
 	case "__dashboard-supervise":
 		return runDashboardSupervise(args[1:])
 	case consumerSpawnVerb: